+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Philosophers Changelog
 
+## currently unfreed mallocs
+- forks
+- threads
+- philo inputs
+
+### v0.2.0 - 11.01.2024
+- created setup function that inits mutexes and starts philo threads
+- created simple thread function that currently just prints some thread input data
+- moved printing function into own file & general cleanup
+- Added epic table visualization
+
+---
+
 ### v0.1.0 - 09.01.2024
 - Added makefile, header, subject, folder structure
 - created logger & atoi utils